Tone down some run-borg errors under chronic
Specifically, failure to acquire the lockfile. This happens regularly
because a previous cron run is still backing things up, and we don't
actually want to fail in this case. But cron still sends me mail
because chronic sees the stderr warning and the nonzero exit code, and
so it sends the output along to cron which triggers the mail.

The old behavior is still probably the right behavior everywhere
*except* under cron, though (i.e. it's what you would want in
interactive use, or if I ever use it somewhere else in a script). So,
what we do is detect when we're running under chronic, and dial down
the severity appropriately.